fix(server): simplify feature flag header parsing
